Completed Project Euler problem #008
Probably would have been better to handle the number variable in a cleaner way than just having it on one endless line, but it worked well enough for the purpose. If it was for a different project I would have stored it in a file and called from that to keep the file clean.